+++ lib/karafka/web/ui/app.rb
@@ -23,7 +23,7 @@ module Karafka
           r.get 'dashboard' do
             @breadcrumbs = false
             controller = Controllers::Dashboard.new(params)
-            render_response controller.index
+            controller.index
           end
 
           r.on 'consumers' do
@@ -34,7 +34,7 @@ module Karafka
             r.get do
               @breadcrumbs = false
               controller = Controllers::Consumers.new(params)
-              render_response controller.index
+              controller.index
             end
           end
 
@@ -50,41 +50,41 @@ module Karafka
 
           r.get 'jobs' do
             controller = Controllers::Jobs.new(params)
-            render_response controller.index
+            controller.index
           end
 
           r.on 'routing' do
             controller = Controllers::Routing.new(params)
 
             r.get String do |topic_id|
-              render_response controller.show(topic_id)
+              controller.show(topic_id)
             end
 
             r.get do
-              render_response controller.index
+              controller.index
             end
           end
 
           r.get 'cluster' do
             controller = Controllers::Cluster.new(params)
-            render_response controller.index
+            controller.index
           end
 
           r.on 'errors' do
             controller = Controllers::Errors.new(params)
 
             r.get Integer do |offset|
-              render_response controller.show(offset)
+              controller.show(offset)
             end
 
             r.get do
-              render_response controller.index
+              controller.index
             end
           end
 
           r.get 'status' do
             controller = Controllers::Status.new(params)
-            render_response controller.show
+            controller.show
           end
         end
       end